Another major rework of 2d shader - now tilts primitives towards viewer.
Before, the primitives were either horizontal or vertical. Now, the primitives
are tilted towards the viewer and the width is adjusted to match the view.
i.e. when looking from above/below the primitive width matches the input
line width and when looking from the side the primitive width matches the
layer height with a smooth(ish) transition between the two.