Case-insensitive visibility checking for printers
In Qt's ListModel, the section headers collect items case-insensitively. However, when checking whether a printer should be visible (when the section header is expanded) was still checked case-sensitively. As a result, it was hiding all printers that did not use the same casing in its manufacturer name, e.g. FlashForge vs. Flashforge.

Fixes #11125.